+++ src/Global.cpp
@@ -732,12 +732,12 @@ void AFXAPI DDX_MyText(CDataExchange* pDX, int nIDC, double& value, double def, 
 
 void AFXAPI DDX_MyText(CDataExchange* pDX, int nIDC, DWORD& value, DWORD def, LPCTSTR pszSuffix)
 {
-	CString strText = FormatString(_T("%u%s"), value, CString(pszSuffix));
+	CString strText = FormatString(_T("%lu%s"), value, CString(pszSuffix));
 	DDX_Text(pDX, nIDC, strText);
 
 	if (pDX->m_bSaveAndValidate)
 	{
-		if (_stscanf(strText, _T("%u"), &value) != 1)
+		if (_stscanf(strText, _T("%lu"), &value) != 1)
 			value = def;
 	}
 }